BAM 900 update HomeBam / Add testVenues
